Generate vehicle ID as u16 again, as we transmit it as f32, which can't hold u32

+++ kissmp-server/src/vehicle/mod.rs
@@ -101,7 +101,7 @@ impl crate::Server {
     }
 
     pub async fn spawn_vehicle(&mut self, client_id: u32, data: VehicleData) {
-        let server_id = rand::random::<u32>();
+        let server_id = rand::random::<u16>();
         let mut data = data.clone();
         data.server_id = server_id;
         data.owner = Some(client_id);
